requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
watchdog